Reduce visibility of artifact constructor, move contexts into model package
