+++ src/main/java/org/tudo/sse/analyses/MavenCentralArtifactAnalysis.java
@@ -6,8 +6,8 @@ import org.tudo.sse.CLIException;
 import org.tudo.sse.model.Artifact;
 import org.tudo.sse.model.ArtifactIdent;
 import org.tudo.sse.model.index.IndexInformation;
-import org.tudo.sse.model.resolution.ArtifactResolutionContext;
-import org.tudo.sse.model.resolution.ResolutionContext;
+import org.tudo.sse.model.ArtifactResolutionContext;
+import org.tudo.sse.model.ResolutionContext;
 import org.tudo.sse.multithreading.ProcessIdentifierMessage;
 import org.tudo.sse.multithreading.WorkloadIsFinalMessage;
 import org.tudo.sse.resolution.ResolverFactory;
