+++ src/main/java/org/tudo/sse/resolution/PomResolver.java
@@ -15,7 +15,7 @@ import org.tudo.sse.model.*;
 import org.tudo.sse.model.pom.License;
 import org.tudo.sse.model.pom.PomInformation;
 import org.tudo.sse.model.pom.RawPomFeatures;
-import org.tudo.sse.model.resolution.ResolutionContext;
+import org.tudo.sse.model.ResolutionContext;
 import org.tudo.sse.resolution.releases.DefaultMavenReleaseListProvider;
 import org.tudo.sse.resolution.releases.IReleaseListProvider;
 import org.tudo.sse.utils.MavenCentralRepository;
